seperate C implementation
